Move LogMapper to the correct package
